Move memory benchmark to a dedicated file
