+++ .github/workflows/benchmark.yml
@@ -57,4 +57,4 @@ jobs:
         uses: CodSpeedHQ/action@f99becdce5e5d51fd556489ebef684f4ecfd6286 # v4.18.5
         with:
           mode: memory
-          run: uv run pytest tests/test_benchmark.py::test_bench_http2_send_large_body --codspeed
+          run: uv run pytest tests/test_benchmark_memory.py --codspeed
